+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Fixed being able to break scripted_sequence by +using friendly NPCs to make them follow player [#200](https://github.com/SamVanheer/halflife-updated/issues/200) (Thanks Oxofemple. for reporting this and FreeSlave for finding the solution)
 * Fixed potential incorrect facing in scripted sequence (Thanks FreeSlave)
 * Made the Linux version link statically to the C++ runtime to help avoid problems when running mods on older systems (Thanks a1ba and FreeSlave)
+* Fixed Egon not stopping its attack animation if the attack button is held down and ammo runs out
 
 ## Changes in V1.0.0 Beta 014